avoid extra looping

+++ src/main/java/chibi/gemmaanalysis/DiffExWithBatchCleanupCli.java
@@ -72,18 +72,24 @@ public class DiffExWithBatchCleanupCli extends ExpressionExperimentManipulatingC
 
             differentialExpressionAnalysisService.thaw( diffAnalyses );
 
-            a: for ( DifferentialExpressionAnalysis existingAnalysis : diffAnalyses ) {
+            for ( DifferentialExpressionAnalysis existingAnalysis : diffAnalyses ) {
 
+                boolean shouldDelete = false;
                 for ( ExpressionAnalysisResultSet resultSet : existingAnalysis.getResultSets() ) {
                     if ( resultSet.getExperimentalFactors().size() > 1 ) {
-                        continue a;
+                        continue;
                     }
                     ExperimentalFactor factor = resultSet.getExperimentalFactors().iterator().next();
                     if ( factor.getName().equals( "batch" ) ) {
+                        shouldDelete = true;
+                        break;
+                    }
+
+                    if ( shouldDelete ) {
                         log.info( "Found analysis with batch factor, Id=" + existingAnalysis.getId() );
                         ds.deleteOldAnalysis( expressionExperiment, existingAnalysis );
-
                     }
+
                 }
 
             }
